requirements.txt
pure-python-adb
pysimplegui
pillow